commiting richton's changes
git-svn-id: https://oss-services.rutgers.edu/svn/public/trunk/SPECS@1884 47757528-202a-0410-98a5-d1d1e9a6c406

+++ php.spec
@@ -13,7 +13,7 @@
 Summary: The PHP scripting language
 Name: php
 Version: %{php_ver}
-Release: 1
+Release: 3
 License: PHP License
 Group: Development/Languages
 Source0: php-%{php_ver}.tar.bz2
@@ -24,7 +24,7 @@ BuildRoot: %{_tmppath}/%{name}-root
 Requires: php-common = %{version}-%{release} apache2-module-php = %{version}-%{release} apache-module-php = %{version}-%{release} 
 BuildRequires: patch freetype2-devel make libmcrypt freetype2 gdbm openldap >= 2.3 openldap-devel >= 2.3 libpng3-devel >= 1.2.8 libjpeg >= 6b-11
 BuildRequires: openssl >= 0.9.7e
-BuildRequires: apache apache-devel = %{apache_ver} apache2 apache2-devel = %{apache2_ver} curl freetds-devel freetds-lib 
+BuildRequires: apache apache-devel = %{apache_ver} apache2 apache2-devel = %{apache2_ver} curl freetds-devel freetds-lib postfix-tls
 
 
 %description
@@ -93,6 +93,11 @@ mv ../imap-2004g ./
 
 
 %build
+
+if [ ! -x /usr/lib/sendmail ];then
+  echo "/usr/lib/sendmail must exist. Symlink to postfix's sendmail, perhaps?"
+  exit 1
+fi
 
 #clean the buildroot
 [ %{buildroot} != "/" ] && [ -d %{buildroot} ] && rm -rf %{buildroot};
